fix : Renaming the vue components
Component names should always be multi-word
