+++ src/components/elements/MyIcons.vue
@@ -1,24 +1,24 @@
 <script setup>
 import { computed } from 'vue'
 
-import Cart from '@/components/icons/ui/Cart.vue'
-import Clock from '@/components/icons/ui/Clock.vue'
-import Phone from '@/components/icons/ui/Phone.vue'
-import Play from '@/components/icons/ui/Play.vue'
-import Localisation from '@/components/icons/ui/Localisation.vue'
-import Search from '@/components/icons/ui/Search.vue'
-import Star from '@/components/icons/ui/Star/Star.vue'
-import ArrowLeft from '@/components/icons/ui/arrows/ArrowLeft.vue'
-import ArrowRight from '@/components/icons/ui/arrows/ArrowRight.vue'
-import Next from '@/components/icons/ui/arrows/Next.vue'
-import Previous from '@/components/icons/ui/arrows/Previous.vue'
-import Food_1 from '@/components/icons/restaurant/Food_1.vue'
-import Food_2 from '@/components/icons/restaurant/Food_2.vue'
-import Delivery from '@/components/icons/restaurant/Delivery.vue'
-import Facebook from '@/components/icons/socials/Facebook.vue'
-import Instagram from '@/components/icons/socials/Instagram.vue'
-import LinkedIn from '@/components/icons/socials/LinkedIn.vue'
-import Twitter from '@/components/icons/socials/Twitter.vue'
+import Cart from '@/components/icons/ui/CartIcon.vue'
+import Clock from '@/components/icons/ui/ClockIcon.vue'
+import Phone from '@/components/icons/ui/PhoneIcon.vue'
+import Play from '@/components/icons/ui/PlayIcon.vue'
+import Localisation from '@/components/icons/ui/LocalisationIcon.vue'
+import Search from '@/components/icons/ui/SearchIcon.vue'
+import Star from '@/components/icons/ui/Star/StarIcon.vue'
+import ArrowLeft from '@/components/icons/ui/arrows/ArrowLeftIcon.vue'
+import ArrowRight from '@/components/icons/ui/arrows/ArrowRightIcon.vue'
+import Next from '@/components/icons/ui/arrows/NextIcon.vue'
+import Previous from '@/components/icons/ui/arrows/PreviousIcon.vue'
+import Food_1 from '@/components/icons/restaurant/Food_1Icon.vue'
+import Food_2 from '@/components/icons/restaurant/Food_2Icon.vue'
+import Delivery from '@/components/icons/restaurant/DeliveryIcon.vue'
+import Facebook from '@/components/icons/socials/FacebookIcon.vue'
+import Instagram from '@/components/icons/socials/InstagramIcon.vue'
+import LinkedIn from '@/components/icons/socials/LinkedInIcon.vue'
+import Twitter from '@/components/icons/socials/TwitterIcon.vue'
 
 const props = defineProps({
   name: String,
