+++ src/components/icons/restaurant/DeliveryIcon.vue
