+++ src/components/icons/restaurant/Food_1Icon.vue
