+++ src/components/icons/restaurant/Food_2Icon.vue
