+++ src/components/icons/socials/LinkedInIcon.vue
