+++ src/components/icons/ui/CartIcon.vue
