+++ src/components/icons/ui/PlayIcon.vue
