+++ src/components/icons/ui/SearchIcon.vue
