+++ src/components/icons/ui/arrows/ArrowLeftIcon.vue
