+++ src/components/icons/ui/arrows/ArrowRightIcon.vue
