+++ src/components/icons/ui/arrows/NextIcon.vue
