+++ src/components/icons/ui/arrows/PreviousIcon.vue
