+++ src/components/icons/ui/star/StarIcon.vue
